+++ example/src/index.tsx
@@ -5,12 +5,8 @@ import { Navigation, Auth, useAuth, Dropdown } from "react-uicomp";
 const Page1 = () => {
   return (
     <div>
-      <Dropdown
-        toggleOnTriggerElementClick
-        dismissOnOutsideClick={false}
-        triggerElement={() => <button>Toggle Menu</button>}
-      >
-        <div>Dropdown Element</div>
+      <Dropdown triggerElement={() => <button>Toggle Menu</button>}>
+        <div style={{ background: "red", width: 150 }}>Dropdown Element</div>
       </Dropdown>
     </div>
   );
