+++ src/ui/index.ts
@@ -1,2 +1,1 @@
-export * from "./common";
-export * from "./hooks";
+export * from "./Dropdown";
